Moved syncNS to its own module
